+++ src/ExpressionPowerTools.Core/Comparisons/ExpressionEquivalency.cs
@@ -20,10 +20,16 @@ namespace ExpressionPowerTools.Core.Comparisons
     public static class ExpressionEquivalency
     {
         /// <summary>
+        /// A lazy proxy to the rules.
+        /// </summary>
+        private static readonly Lazy<IExpressionComparisonRuleProvider> Provider =
+            ServiceHost.GetLazyService<IExpressionComparisonRuleProvider>();
+
+        /// <summary>
         /// Gets the configured rule set.
         /// </summary>
         private static IExpressionComparisonRuleProvider Rules =>
-            ServiceHost.GetService<IExpressionComparisonRuleProvider>();
+            Provider.Value;
 
         /// <summary>
         /// Determine if a <see cref="Type"/> is equivalent to another type.