+++ src/ExpressionPowerTools.Serialization/Rules/RulesEngine.Config.cs
@@ -50,6 +50,7 @@ namespace ExpressionPowerTools.Serialization.Rules
             var ctorSelector = new MemberSelector<ConstructorInfo>();
             selector(ctorSelector);
             memberInfo = ctorSelector.Member;
+            compiled = false;
             return this;
         }
 
@@ -64,6 +65,7 @@ namespace ExpressionPowerTools.Serialization.Rules
             var fieldSelector = new MemberSelector<FieldInfo>();
             selector(fieldSelector);
             memberInfo = fieldSelector.Member;
+            compiled = false;
             return this;
         }
 
@@ -78,6 +80,7 @@ namespace ExpressionPowerTools.Serialization.Rules
             var methodSelector = new MemberSelector<MethodInfo>();
             selector(methodSelector);
             memberInfo = methodSelector.Member;
+            compiled = false;
             return this;
         }
 
@@ -92,6 +95,7 @@ namespace ExpressionPowerTools.Serialization.Rules
             var propSelector = new MemberSelector<PropertyInfo>();
             selector(propSelector);
             memberInfo = propSelector.Member;
+            compiled = false;
             return this;
         }
 
@@ -104,6 +108,7 @@ namespace ExpressionPowerTools.Serialization.Rules
         {
             CheckMemberInfo(shouldExist: false);
             memberInfo = new[] { type };
+            compiled = false;
             return this;
         }